// lib/api/generated/models/utils-content-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const UtilsContentType = {
    ContentTypeHumanGenerated: 'human-generated',
    ContentTypeAIGenerated: 'ai-generated'
} as const;

export type UtilsContentType = typeof UtilsContentType[keyof typeof UtilsContentType];